fix(runtime): harden failure attribution policy
